updated for get_files_hash_value
