+++ tmVhdlProducer/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "2.17.2"
+__version__ = "2.18.0"
