+++ tmVhdlProducer/vhdlhelper.py
@@ -53,9 +53,7 @@ import tmGrammar  # import after tmEventSetup
 
 from . import algodist
 from . import __version__
-
-import hashlib
-from glob import glob
+from .constants import get_files_hash_value
 
 # -----------------------------------------------------------------------------
 #  Precompiled regular expressions
@@ -254,31 +252,6 @@ def bx_encode_4_array(value: int) -> str:
     p2 is array index 0, p1 is array index 1, and so on.
     """
     return format([2, 1, 0, -1, -2].index(value), 'd')
-
-def get_files_hash_value(dir_path):
-    """Calculate hash value of the content of all .py and .vhd files in 'dir_path'.
-    """
-    py_files = dir_path+"/**/*.py"
-    vhd_files = dir_path+"/**/*.vhd"
-
-    x = hashlib.sha256()
-
-    filename = []
-
-    for filename_py in glob(py_files, recursive=True):
-        filename.append(filename_py)
-    for filename_vhd in glob(vhd_files, recursive=True):
-        filename.append(filename_vhd)
-    for i in range(0, len(filename)):
-        with open(filename[i], 'rb') as f:
-            while True:
-                # Reading is buffered, so we can read smaller chunks.
-                chunk = f.read(x.block_size)
-                if not chunk:
-                    break
-                x.update(chunk)
-
-    return x.hexdigest()
 
 # -----------------------------------------------------------------------------
 #  Factories
